# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Avelorn",
    "summary": "The unit and army database for Warhammer: The Old World.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/units/{slug}": {
      "get": {
        "summary": "Read one datasheet",
        "description": "Read a datasheet in full: its profiles, equipment, rules, and options.\n\nReturns:\n    The datasheet, its troop-type profile resolved and each printed rule\n    name carrying the entry it resolves to.\n\nRaises:\n    HTTPException: 404, when no datasheet carries the slug.",
        "operationId": "read_unit_units__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/muster": {
      "post": {
        "summary": "Cost and equip one block of an army list",
        "description": "Size and equip a datasheet, and derive what the block costs.\n\nSays nothing about whether a list of these is legal -- army composition\nis not modelled yet. This costs one block and refuses one the datasheet\ndoes not allow. A ``frontage`` re-forms the block that many models wide,\nwhich changes the footprint it stands on and nothing about its cost.\n\nReturns:\n    The block, its points and effective loadout derived.\n\nRaises:\n    HTTPException: 404, when no datasheet carries the slug; 422, when the\n        datasheet does not allow the size or the options asked for.",
        "operationId": "muster_muster_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Muster"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusteredUnit"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fight": {
      "post": {
        "summary": "Resolve one round of close combat between two units",
        "description": "Deploy two units, fight one round, and score it.\n\nOne round: both sides strike in Initiative order, the Wounds tally into a\ncombat result, and the loser takes its Break test. What a round does not\ncover is the rest of the engagement \u2014 a pursuit, a second round, and the\nStand & Shoot a charge would be met with, which needs the Movement phase's\ncharge sequence rather than a charge recorded on the charger.\n\nA side the corpus cannot field is refused before any dice are walked: an\nunknown slug is a 404, and a size, option or weapon the datasheet does not\nallow is a 422 naming which side asked for it.\n\nReturns:\n    The round resolved: each side's casualty distribution and Break-test\n    outcomes, who won, and every rule the engine held without applying.",
        "operationId": "fight_fight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Fight"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FightReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/volley": {
      "post": {
        "summary": "Resolve one volley of shooting, and the panic it causes",
        "description": "Shoot one unit at another and resolve the panic its casualties cause.\n\n``distance`` is in inches; ``moved`` says the shooter moved this turn.\n``battle_strength`` is the target's size at the start of the battle, which\ngoverns the Fall Back or Flee split; it defaults to the size it is shot at.\n\nA side the corpus cannot field is refused before any dice are walked: an\nunknown slug is a 404, and a size, option or weapon the datasheet does not\nallow is a 422 naming which side asked for it. The shooter must carry\nsomething with a missile profile.\n\nReturns:\n    The volley resolved: the scores each roll needs, the wound and casualty\n    distributions, what the target's nerve does, and every rule the volley\n    holds without applying.",
        "operationId": "volley_volley_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Volley"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolleyReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/weapons/{slug}": {
      "get": {
        "summary": "Read one weapon entry",
        "description": "Read a weapon entry in full: its profiles, its rules, its restrictions.\n\nReturns:\n    The entry, each profile's printed rule names carrying what they resolve\n    to.\n\nRaises:\n    HTTPException: 404, when no entry carries the slug.",
        "operationId": "read_weapon_weapons__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeaponDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/armour/{slug}": {
      "get": {
        "summary": "Read one armour entry",
        "description": "Read one armour entry: its armour value, and what it leaves out.\n\nReturns:\n    The entry.\n\nRaises:\n    HTTPException: 404, when no entry carries the slug.",
        "operationId": "read_armour_armour__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Armour"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/rules/{slug}": {
      "get": {
        "summary": "Read one rule entry",
        "description": "Read a rule entry in full: its text, its effects, and what it leaves out.\n\nReturns:\n    The rule entry.\n\nRaises:\n    HTTPException: 404, when no entry carries the slug. A rule the corpus\n        prints without an entry has none to read; ``/coverage``\n        names those.",
        "operationId": "read_rule_rules__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rule"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Fight": {
        "properties": {
          "a": {
            "$ref": "#/components/schemas/Deployment"
          },
          "b": {
            "$ref": "#/components/schemas/Deployment"
          },
          "charge": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChargedBy"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "a",
          "b"
        ],
        "title": "Fight",
        "description": "Two sides meeting in close combat, and the charge that brought them together."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Muster": {
        "properties": {
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "frontage": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Frontage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "unit",
          "size"
        ],
        "title": "Muster",
        "description": "What a caller asks to field: a datasheet, a model count, and options by name."
      },
      "MusteredUnit": {
        "properties": {
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "equipment": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Equipment"
          },
          "weapons": {
            "items": {
              "$ref": "#/components/schemas/Wieldable"
            },
            "type": "array",
            "title": "Weapons"
          },
          "special_rules": {
            "items": {
              "$ref": "#/components/schemas/Reference"
            },
            "type": "array",
            "title": "Special Rules"
          },
          "footprint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Footprint"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "unit",
          "name",
          "size",
          "options",
          "points",
          "equipment",
          "weapons",
          "special_rules",
          "footprint"
        ],
        "title": "MusteredUnit",
        "description": "A block of an army list: a datasheet sized and equipped, and what it costs.\n\nThe view over :class:`~avelorn.tow.muster.Complement`. It carries the\ndatasheet by slug rather than whole, because a list is read as a list --\na caller wanting the profiles follows ``unit`` to the datasheet route.\n``equipment`` and ``special_rules`` are the effective ones, the chosen\noptions' adds and removes already folded in, so a block says what the\nmodels actually carry rather than what the datasheet offered. ``weapons``\nnarrows the equipment to the weapons among it, each saying whether it can\nbe used in close combat -- what a caller naming a weapon chooses from.\n``footprint`` is the table space the block takes, at the frontage asked for\nor the datasheet's default, which a caller drawing it needs and cannot\nderive from a slug."
      },
...
      "UnitDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "points": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Points"
          },
          "unit_size": {
            "$ref": "#/components/schemas/UnitSize"
          },
          "troop_type": {
            "$ref": "#/components/schemas/TroopType"
          },
          "troop_type_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TroopTypeProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "base_size": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BaseSize"
              },
              {
                "type": "null"
              }
            ]
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/Profile"
            },
            "type": "array",
            "minItems": 1,
            "title": "Profiles"
          },
          "equipment": {
            "items": {
              "$ref": "#/components/schemas/Reference"
            },
            "type": "array",
            "title": "Equipment"
          },
          "special_rules": {
            "items": {
              "$ref": "#/components/schemas/Reference"
            },
            "type": "array",
            "title": "Special Rules"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/OptionDetail"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "points",
          "unit_size",
          "troop_type",
          "profiles",
          "equipment",
          "special_rules",
          "options"
        ],
        "title": "UnitDetail",
        "description": "A datasheet as reading one shows it: the entry, every printed name resolved.\n\nEverything :class:`~avelorn.tow.schema.unit.Unit` prints, except that its\nequipment, its special rules and its options' rules arrive as\n:class:`Reference`. Resolving on the way out is what keeps a caller from\nre-deriving it: a printed name does not become a slug by slugifying it, nor\na reference a name by reading it."
      },
      "UnitSize": {
        "properties": {
          "min": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "min"
        ],
        "title": "UnitSize",
        "description": "Allowed model count for a unit."
      },
...
      "Volley": {
        "properties": {
          "shooter": {
            "$ref": "#/components/schemas/Deployment"
          },
          "target": {
            "$ref": "#/components/schemas/Deployment"
          },
          "distance": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Distance"
          },
          "moved": {
            "type": "boolean",
            "title": "Moved",
            "default": false
          },
          "battle_strength": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Battle Strength"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "shooter",
          "target",
          "distance"
        ],
        "title": "Volley",
        "description": "One unit shooting another, and what the shot has to travel through."
      },
...
    }
  }
}